zkevm prototype (#52)
Forkable zkevm implementation.

The only tricky thing was how to deal with the state updates.
There are 4 state updates(https://zkevm.polygon.technology/docs/protocol/state-management)
sequencer confirmation(for us unimportant)
sequenced(after the sequencer submitted payload to L1). We will not consider this for the forking process
verified(after a zk proof has been submitted)
consolidated(after a period called pendingStateTimeout - if the proof was given by someone random - or after verified by the trusted sequencer).

I decided that I want to only post consolidated stateRoots into the new forks. This gives the guarantees similar to the original implementation, but it might be that the verified state is not considered, in a forking situations, if it is not yet consolidated.

This might not give the nicest UX during forking, but since forking is not happening frequently, we should go for it, as it improves the security.