Removing UUPS in order to reduce the size of the contracts (#56)
I thought in the past that UUPS proxy structure is ideal for us. I have not changed my mind, but I figured out that it bloats the contract size too much. UUPS contract was 6kb on its own. And with contracts over 30kB in size, we could not deploy them to mainnet.

This PR removes UUPS structure and uses a normal eip1967 Proxy. Ie. the upgrading logic will be removed from the original contact and we can later push it into the proxies itself.
Also, I am using two more libraries to reduce code size.