Forkaware structure (#74)
On L2, I think many contracts need to wait after a fork for the first message that will come from the bridge. E.g. our adjudication framework would have to wait for a message which court is removed before it starts to forward any new information of the oracles/courts. Then I think the following contract will be quite helpful.