E2e test with real verification after forking (#133)
/**

This test simulates the first proof after a fork
The stateRoot and the globalExitRoot of the old state was calculated with an old chainID: 1000
The new chainID is 1001. The test shows that the newly generated proof is valid
on the real verifier contract. To show this a polygonZkEVM contract is set up
with the old state root, as it is done in a fork scenario via forkingManager contract.
After this a new batch is sequenced and the proof is verified.
The old state root was generated with the zkevm state tool from common-js. See this fork:
https://github.com/RealityETH/zkevm-commonjs/pull/1/files
and especially this line:
https://github.com/RealityETH/zkevm-commonjs/pull/1/files#diff-4d0a534c9be88e9cfd08ea7c2e6d412b9ad1d6501db1900339f19c02fb689922R147
It took the genesis state from our deployment script and the first two transactions are
some transfers on the chain.
Note that we run this "deterministic test" without a fork introduced by the forkingManager,
as this would require us to generate a proof for specific inputs (stateHash, globalExitRoot, etc.)
that would require us to prover each new test-run (unless one is able to set up the test perfectly deterministic as well).
This would be additional work and is not necessary to verify that we can
continue to use the old state root and globalExitRoot with a new chainID - as it is done in this test
*/

test can be run via:
npx hardhat test test/e2e/forking.t.js