No remapping for hardhat (#142)
To test it, run npx hardhat compile and then one more time. Caching should hit in, then.

Background: We want to get rid of the remapping usages in hardhat, as it has bugs. The solution is: we just install the deps twice: once via git submodule for foundry and once via npm for hardhat.
We use in the contract only the npm packages, but allow foundry to remap to submodules for its purposes.

Longterm we wanna switch to hardhat tests and away from foundry