e2e test for withdrawals L2-initiate -> forking -> L1-finalize (#143)
This test simulates the withdrawal process after a fork
At chainstart the genesis is similar to the output of the genesis script form this repo
In the next block a test account makes a deposit of 10 native tokens (forkonomic tokens)
on L2 into the bridge. The transaction was created with the script despoitForkonomicTokenIntoBridgeInL2.js.
Then this state advancement with the deposit is verified with a snark proof and submitted. See this PR for the actual state advancement:
preparing proof inputs for withdrawal e2e test zkevm-commonjs#2
After that the fork is initiated and executed. We then test the withdrawal process.